2nd time committing end of class changes
