+++ db/index.js
@@ -51,7 +51,7 @@ async function createMeat({
     } = await client.query(
       `
       INSERT INTO meat(species,style,description,flavor,weight,price) 
-      VALUES($1, $2, $3, $4,$5,$6) 
+      VALUES($1, $2, $3, $4, $5, $6) 
       RETURNING *;
     `,
       [species, style, description, flavor, weight, price]
@@ -63,119 +63,84 @@ async function createMeat({
   }
 }
 
+async function createOrder({ user_id, fufilled }) {
+  try {
+    const {
+      rows: [user],
+    } = await client.query(
+      `
+      INSERT INTO orders(user_id, fufilled) 
+      VALUES($1, $2)
+      RETURNING *;
+    `,
+      [user_id, fufilled]
+    );
+    return user;
+  } catch (error) {
+    throw error;
+  }
+}
+
+async function createUserMeats({ meat_id, user_id, order_id }) {
+  try {
+    const {
+      rows: [user],
+    } = await client.query(
+      `
+      INSERT INTO user_meats(meat_id, user_id, order_id) 
+      VALUES($1, $2, $3)
+      RETURNING *;
+    `,
+      [meat_id, user_id, order_id]
+    );
+    return user;
+  } catch (error) {
+    throw error;
+  }
+}
+
 // ---------------------------------
-async function dropTables() {
-  try {
-    console.log("Dropping Tables");
-    await client.query(`
-    DROP TABLE IF EXISTS user_meats;
-    DROP TABLE IF EXISTS meat;
-    DROP TABLE IF EXISTS orders;
-    DROP TABLE IF EXISTS users;`);
-  } catch (error) {
-    console.log("Error Dropping Tables");
-    throw error;
-  }
-}
 
-async function createTables() {
+async function getAllMeats() {
   try {
-    console.log("Creating Tables");
-
-    await client.query(`
-    CREATE TABLE users (
-      id SERIAL PRIMARY KEY,
-      email varchar(255) UNIQUE NOT NULL,
-      password varchar(255) NOT NULL
+    const response = await client.query(
+      `
+      SELECT * FROM meat;
+    `
     );
-    
-    CREATE TABLE meat (
-      id SERIAL PRIMARY KEY,
-      species varchar (255) NOT NULL,
-      style varchar (255) NOT NULL,
-      description varchar(255),
-      flavor varchar (255) NOT NULL,
-      weight DECIMAL NOT NULL,
-      price DECIMAL NOT NULL
-    );
-    
-    CREATE TABLE orders (
-      id SERIAL PRIMARY KEY,
-      user_id INTEGER REFERENCES users(id),
-      fufilled BOOLEAN DEFAULT false
-    );
-    
-    CREATE TABLE user_meats (
-      id SERIAL PRIMARY KEY,
-      meat_id INTEGER REFERENCES meat(id),
-      user_id INTEGER REFERENCES users(id)
-    );`);
+    return response.rows;
   } catch (error) {
-    console.log("Error Creating Tables");
-    throw error;
-  }
-}
-async function createInitialUsers() {
-  try {
-    console.log("Creating Users");
-    await createUser({
-      email: "hotmeat1@hotmail.com",
-      password: "Greasy1",
-    });
-    await createUser({
-      email: "lilsmokey@bigdogzonly.com",
-      password: "Greasy2",
-    });
-    await createUser({
-      email: "papaSausage@gmail.com",
-      password: "Carved1",
-    });
-  } catch (error) {
-    console.log("Error Creating Users");
     throw error;
   }
 }
 
-async function createInitialMeats() {
-  console.log("Making Meats");
+async function getAllOrders() {
   try {
-    await createMeat({
-      species: "Cow",
-      style: "Ground Beef",
-      description: "tender, low fat, grass fed",
-      flavor: "plain",
-      weight: 1,
-      price: 5,
-    });
-    await createMeat({
-      species: "Gnarwall",
-      style: "Filet",
-      description:
-        "Angled from the northern most part of the north sea, sliced and diced for your pallet pleasure",
-      flavor: "delicious",
-      weight: 0.25,
-      price: 700,
-    });
-    await createMeat({
-      species: "Bison",
-      style: "Burger",
-      description:
-        "Hunted by Buffalo Bill himself, Has been marinating for 100 years",
-      flavor: "American",
-      weight: 20,
-      price: 1000,
-    });
+    const {
+      rows: [orders],
+    } = await client.query(
+      `
+      SELECT * FROM orders;
+    `
+    );
+    return orders;
   } catch (error) {
-    console.log("Error making Meat");
     throw error;
   }
 }
 
-dropTables();
-createTables();
-createInitialUsers();
-createInitialMeats();
+getAllMeats().then((value) => {
+  console.log(value);
+});
 
 client.connect();
 
-module.exports = client;
+module.exports = {
+  client,
+  createMeat,
+  createOrder,
+  createUser,
+  createUserMeats,
+  getAllMeats,
+  getAllOrders,
+};
